+++ debug.log
@@ -1,0 +1,8 @@
+[0417/192431.411:ERROR:tcp_listen_socket.cc(76)] Could not bind socket to 127.0.0.1:5859
+[0417/192431.412:ERROR:node_debugger.cc(87)] Cannot start debugger server
+[0417/192432.578:ERROR:tcp_listen_socket.cc(76)] Could not bind socket to 127.0.0.1:5860
+[0417/192432.581:ERROR:node_debugger.cc(87)] Cannot start debugger server
+[0417/192850.371:ERROR:tcp_listen_socket.cc(76)] Could not bind socket to 127.0.0.1:5859
+[0417/192850.372:ERROR:node_debugger.cc(87)] Cannot start debugger server
+[0417/192851.897:ERROR:tcp_listen_socket.cc(76)] Could not bind socket to 127.0.0.1:5860
+[0417/192851.898:ERROR:node_debugger.cc(87)] Cannot start debugger server
